third_party/WebKit/LayoutTests/storage/indexeddb/idbcursor-delete-exception-order.html

<!DOCTYPE html>
<title>IndexedDB: IDBCursor delete() Exception Ordering</title>
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbcursor-delete">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/testharness-helpers.js"></script>

<script>

indexeddb_test(
    function(t, db) {
        var s = db.createObjectStore('s');
        s.put('value', 'key');
    },
    function(t, db) {
        var s = db.transaction('s', 'readonly').objectStore('s');
        var r = s.openCursor();
        r.onsuccess = t.step_func(function() {
            r.onsuccess = null;
            var cursor = r.result;
            setTimeout(t.step_func(function() {
                assert_throws('TransactionInactiveError', function() {
                    cursor.delete();
                }, '"Transaction inactive" check (TransactionInactivError) ' +
                   'should precede "read only" check (ReadOnlyError)');
                t.done();
            }), 0);
        });
    },
    'IDBCursor.delete exception order: TransactionInactiveError vs. ReadOnlyError'
);

indexeddb_test(
    function(t, db) {
        var s = db.createObjectStore('s');
        s.put('value', 'key');
    },
    function(t, db) {
        var s = db.transaction('s', 'readonly').objectStore('s');
        var r = s.openCursor();
        r.onsuccess = t.step_func(function() {
            r.onsuccess = null;
            var cursor = r.result;
            cursor.continue();
            assert_throws('ReadOnlyError', function() {
                cursor.delete();
            }, '"Read only" check (ReadOnlyError) should precede ' +
               '"got value flag" (InvalidStateError) check');
            t.done();
        });
    },
    'IDBCursor.delete exception order: ReadOnlyError vs. InvalidStateError #1'
);

indexeddb_test(
    function(t, db) {
        var s = db.createObjectStore('s');
        s.put('value', 'key');
    },
    function(t, db) {
        var s = db.transaction('s', 'readonly').objectStore('s');
        var r = s.openKeyCursor();
        r.onsuccess = t.step_func(function() {
            r.onsuccess = null;
            var cursor = r.result;
            cursor.continue();
            assert_throws('ReadOnlyError', function() {
                cursor.delete();
            }, '"Read only" check (ReadOnlyError) should precede ' +
               '"key only flag" (InvalidStateError) check');
            t.done();
        });
    },
    'IDBCursor.delete exception order: ReadOnlyError vs. InvalidStateError #2'
);

</script>
